changed Person's Model, shifted original code to , refactored all references to old Model and ModelManager
