+++ src/main/java/oldCode/Model.java
@@ -1,8 +1,9 @@
-package seedu.address.model;
+package oldCode;
 
 import java.util.function.Predicate;
 
 import javafx.collections.ObservableList;
+import seedu.address.model.ReadOnlyAddressBook;
 import seedu.address.model.person.Person;
 
 /**
