+++ src/main/java/seedu/address/MainApp.java
@@ -29,8 +29,8 @@ import seedu.address.controller.LoginController;
 import seedu.address.logic.Logic;
 import seedu.address.logic.LogicManager;
 import seedu.address.model.AddressBook;
-import seedu.address.model.Model;
-import seedu.address.model.ModelManager;
+import oldCode.Model;
+import oldCode.ModelManager;
 import seedu.address.model.ReadOnlyAddressBook;
 import seedu.address.model.UserPrefs;
 import seedu.address.model.util.SampleDataUtil;
