+++ src/main/java/seedu/address/logic/commands/BatchCommand.java
@@ -8,7 +8,7 @@ import seedu.address.logic.CommandHistory;
 import seedu.address.logic.commands.exceptions.CommandException;
 
 import seedu.address.model.IngredientModel;
-import seedu.address.model.Model;
+import oldCode.Model;
