+++ src/main/java/seedu/address/logic/commands/ClearCommand.java
@@ -4,7 +4,7 @@ import static java.util.Objects.requireNonNull;
 
 import seedu.address.logic.CommandHistory;
 import seedu.address.model.AddressBook;
-import seedu.address.model.Model;
+import oldCode.Model;
 
 /**
  * Clears the address book.
