+++ src/main/java/seedu/address/logic/commands/Command.java
@@ -2,7 +2,7 @@ package seedu.address.logic.commands;
 
 import seedu.address.logic.CommandHistory;
 import seedu.address.logic.commands.exceptions.CommandException;
-import seedu.address.model.Model;
+import oldCode.Model;
 
 /**
  * Represents a command with hidden internal logic and the ability to be executed.
