+++ src/main/java/seedu/address/logic/commands/DeleteCommand.java
@@ -8,7 +8,7 @@ import seedu.address.commons.core.Messages;
 import seedu.address.commons.core.index.Index;
 import seedu.address.logic.CommandHistory;
 import seedu.address.logic.commands.exceptions.CommandException;
-import seedu.address.model.Model;
+import oldCode.Model;
 import seedu.address.model.person.Person;
 
 /**
