+++ src/main/java/seedu/address/logic/commands/ExitCommand.java
@@ -3,7 +3,7 @@ package seedu.address.logic.commands;
 import seedu.address.commons.core.EventsCenter;
 import seedu.address.commons.events.ui.ExitAppRequestEvent;
 import seedu.address.logic.CommandHistory;
-import seedu.address.model.Model;
+import oldCode.Model;
 
 /**
  * Terminates the program.
