+++ src/main/java/seedu/address/logic/commands/FindCommand.java
@@ -4,7 +4,7 @@ import static java.util.Objects.requireNonNull;
 
 import seedu.address.commons.core.Messages;
 import seedu.address.logic.CommandHistory;
-import seedu.address.model.Model;
+import oldCode.Model;
 import seedu.address.model.person.NameContainsKeywordsPredicate;
 
 /**
