+++ src/main/java/seedu/address/logic/commands/HelpCommand.java
@@ -3,7 +3,7 @@ package seedu.address.logic.commands;
 import seedu.address.commons.core.EventsCenter;
 import seedu.address.commons.events.ui.ShowHelpRequestEvent;
 import seedu.address.logic.CommandHistory;
-import seedu.address.model.Model;
+import oldCode.Model;
 
 /**
  * Format full help instructions for every command for display.
