+++ src/main/java/seedu/address/logic/commands/ListCommand.java
@@ -1,10 +1,10 @@
 package seedu.address.logic.commands;
 
 import static java.util.Objects.requireNonNull;
-import static seedu.address.model.Model.PREDICATE_SHOW_ALL_PERSONS;
+import static oldCode.Model.PREDICATE_SHOW_ALL_PERSONS;
 
 import seedu.address.logic.CommandHistory;
-import seedu.address.model.Model;
+import oldCode.Model;
 
 /**
  * Lists all persons in the address book to the user.
