+++ src/test/java/seedu/address/logic/commands/UndoCommandTest.java
@@ -9,8 +9,8 @@ import org.junit.Before;
 import org.junit.Test;
 
 import seedu.address.logic.CommandHistory;
-import seedu.address.model.Model;
-import seedu.address.model.ModelManager;
+import oldCode.Model;
+import oldCode.ModelManager;
 import seedu.address.model.UserPrefs;
 
 public class UndoCommandTest {
